requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg>=0.29.0
openai>=1.12.0,<2.0.0
python-dotenv==1.0.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1  # For TestClient
redis>=5.0.1
crewai==0.11.0
langchain==0.1.6
duckduckgo-search==4.4.3
setuptools>=75.8.0
gpt4all>=2.0.0  # Added for local LLM support
